+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ## [Unreleased]
 ### Changed
 - url attribute to IssuingToken resource
+- installmentCount attribute to IssuingPurchase resource
 
 ## [0.16.0] - 2026-06-01
 ### Added